+++ default.nix
@@ -31,8 +31,8 @@
 
   yosys ? import ./nix/yosys.nix { inherit pkgs; },
   
-  volare-rev ? "cb90dc49bf6135bab117a69a8acdc24e3d510f6d",
-  volare-sha256 ? "sha256-hnWX59Ha54xnhHKvCvw7HT1KmeZYS7xUXUCmgp+0YSo=",
+  volare-rev ? "2775640b422afa3c1c0016c4a46da5fa23026a93",
+  volare-sha256 ? "sha256-a7I4Tg6tzFGFo3u2zQIKScboOX4OFO4FdqBKbcxEEl0=",
   volare ? let src = pkgs.fetchFromGitHub {
     owner = "efabless";
     repo = "volare";