+++ openlane/scripts/magic/extract_spice.tcl
@@ -38,6 +38,12 @@ if { ! $::env(MAGIC_NO_EXT_UNIQUE) } {
 extract
 
 ext2spice lvs
+
+# For designs where more than one top-level pin is connected to the same net
+if { $::env(MAGIC_EXT_SHORT_RESISTOR) } {
+    ext2spice short resistor
+}
+
 ext2spice -o $netlist $::env(DESIGN_NAME).ext
 feedback save $feedback_file
 # exec cp $::env(DESIGN_NAME).spice $::env(signoff_results)/$::env(DESIGN_NAME).spice