+++ test/config/test_variable.py
@@ -19,9 +19,6 @@ from pyfakefs.fake_filesystem_unittest import Patcher
 from typing import Dict, List, Literal, Optional, Tuple, Type, Union
 
 
-from .variable import Variable
-
-
 @pytest.fixture()
 def _mock_fs():
     with Patcher() as patcher:
@@ -33,7 +30,7 @@ def _mock_fs():
 
 
 def test_macro_validation():
-    from . import Macro
+    from openlane.config import Macro
 
     with pytest.raises(ValueError, match="at least one GDSII file"):
         Macro(gds=[], lef=["test"])
@@ -43,7 +40,7 @@ def test_macro_validation():
 
 
 def test_is_optional():
-    from .variable import is_optional
+    from openlane.config.variable import is_optional
 
     assert is_optional(int) is False, "is_optional false positive"
     assert is_optional(Optional[int]) is True, "is_optional false negative"
@@ -56,7 +53,7 @@ def test_is_optional():
 
 
 def test_some_of():
-    from .variable import some_of
+    from openlane.config.variable import some_of
 
     assert some_of(int) == int, "some_of changed the type of a non-option type"
     assert (
@@ -75,7 +72,7 @@ def test_some_of():
 
 
 def test_variable_construction():
-    from . import Variable
+    from openlane.config import Variable
 
     variable = Variable(
         "EXAMPLE",
@@ -109,7 +106,8 @@ def test_variable_construction():
 
 @pytest.fixture()
 def variable():
-    from .variable import Path, Variable
+    from openlane.common import Path
+    from openlane.config import Variable
 
     return Variable(
         "EXAMPLE",
@@ -120,8 +118,8 @@ def variable():
 
 
 @pytest.mark.usefixtures("_mock_fs")
-def test_compile(variable: Variable):
-    from ..common import GenericDict
+def test_compile(variable):
+    from openlane.common import GenericDict
 
     valid_input = GenericDict({"EXAMPLE": ["/cwd/a", "/cwd/b"]})
     warning_list = []
@@ -138,7 +136,8 @@ def test_compile(variable: Variable):
 
 
 def test_compile_dataclass():
-    from ..common import GenericDict
+    from openlane.common import GenericDict
+    from openlane.config import Variable
 
     @dataclass
     class MyClass:
@@ -155,7 +154,8 @@ def test_compile_dataclass():
 
 
 def test_compile_required():
-    from ..common import GenericDict
+    from openlane.common import GenericDict
+    from openlane.config import Variable
 
     variable = Variable("MY_VARIABLE", int, description="x")
 
@@ -167,8 +167,8 @@ def test_compile_required():
 
 
 @pytest.mark.usefixtures("_mock_fs")
-def test_compile_deprecated(variable: Variable):
-    from ..common import GenericDict
+def test_compile_deprecated(variable):
+    from openlane.common import GenericDict
 
     deprecated_valid_input = GenericDict(
         {
@@ -195,13 +195,15 @@ def test_compile_deprecated(variable: Variable):
 
 @pytest.fixture()
 def test_enum():
-    from ..common import StringEnum
+    from openlane.common import StringEnum
 
     return StringEnum("x", ["AValue", "AnotherValue"])
 
 
 @pytest.fixture()
 def variable_set(variable, test_enum):
+    from openlane.config import Variable
+
     return [
         variable,
         Variable(
@@ -259,8 +261,8 @@ def variable_set(variable, test_enum):
 
 
 @pytest.mark.usefixtures("_mock_fs")
-def test_compile_invalid(variable_set: List[Variable]):
-    from ..common import GenericDict
+def test_compile_invalid(variable_set: list):
+    from openlane.common import GenericDict
 
     invalid_input = GenericDict(
         {
@@ -295,8 +297,8 @@ def test_compile_invalid(variable_set: List[Variable]):
 
 
 @pytest.mark.usefixtures("_mock_fs")
-def test_compile_permissive(variable_set: List[Variable], test_enum: Type):
-    from ..common import GenericDict
+def test_compile_permissive(variable_set: list, test_enum: Type):
+    from openlane.common import GenericDict
 
     permissive_valid_input = GenericDict(
         {